chore: move demo to separate file
