+++ test/jsonFormatArgs.spec.js
@@ -42,31 +42,3 @@ describe('jsonFormatArgs', () => {
     });
   });
 });
-
-// TODO convert these into tests
-/* eslint-disable */
-export function test() {
-  const debug = require('debug');
-  debug.formatArgs = jsonFormatArgs;
-
-  const firstLogger = debug('json:direct');
-  firstLogger('logger, simple');
-  firstLogger('logger, attachments', 'attached');
-  firstLogger('logger, formatting args %%s, %s', 'percent-s');
-
-  firstLogger('logger, recursive arg with %%j, %j', x);
-
-
-  firstLogger('logger, many args, %s', 'inline', 'attached');
-
-  const secondDebug = require('debug');
-  const secondLogger = secondDebug('json:other');
-
-  secondLogger('logger, 0 args');
-}
-
-// module.exports = jsonFormatArgs;
-// console.log(debug.formatters.s);
-// test();
-
-/* eslint-enable */
